+++ suit5/templatetags/suit_menu.py
@@ -402,7 +402,7 @@ class Menu(object):
         self.fill_keys(app, keys)
 
     def ensure_model_keys(self, model):
-        keys = ['label', 'url', 'permissions', 'is_active', 'blank']
+        keys = ['label', 'url', 'icon', 'permissions', 'is_active', 'blank']
         self.fill_keys(model, keys)
 
     def fill_keys(self, dict, keys):